Hello there,
I've added an initial support for rendering template partials, based on mustache partial interpolation with `{{>partial_name}}``

```js

var template = 'Hello {{=it.name}}, you are {{=it.details.age}} years old! {{> menu }}';

var partials = {
    menu : "<p>Hello again {{= it.name}} I still believe you are {{= it.details.age}} right? {{> banner}}</p>",
    banner : "Hey its me! Banner"
};

var data  = {
    name: 'John', 
    details : {age : 27}
};


doT.template(template, null, partials)(data)

//outputs
/*
    Hello John, you are 27 years old! 
    Hello again John I still believe you are 27 right? Hey its me! Banner
*/

```

As you can see the `partials` declaration is being passed in the same time as the template, not data, this is because I believe partials are directly related with the template itself since they can also be classified as `sub-templates`. It is being passed as the 3rd `defs` attribute now only because this one  wasn't being used before. The 3rd attribute could also be passed as an object witch contains the `defl` definition and another one for partials like `{defl : ..., partials : { ... })`

I hope this make sense :)

The while loop existis because it will expand the whole template and partials inside partials and then continue with the process.

I am really not an expert on performance, and I wouldn't like break the beauty of your engine which is performance/speed, so please let me know if this doesn't work the way it should.

Thanks