Adds an engine attribute for easy identification
I am working on an application server and the ability to identify
and handle various templating engines is important. Right now we
are handling detection of doT by looking at the shape of the
template engine. Would be nice to have a single field to check like
handlebars has, for instance.

This adds an engine attribute so that it will be easy to detect in
the future.